Fixes with writing results

+++ benchmarks/benchmark_rpq/test_benchmark_rpq.py
@@ -29,11 +29,12 @@ def test_benchmark_rpq(impl, graph, regex):
         , 'Time (in microseconds)'
     ]
 
-    with open(result_file_path, mode='w+', newline='\n') as f:
-        csv_writer = csv.writer(f, delimiter=',', quoting=csv.QUOTE_NONNUMERIC, escapechar=' ')
-        csv_writer.writerow(headers)
+    if not os.path.exists(result_file_path):
+        with open(result_file_path, mode='w+', newline='\n') as f:
+            csv_writer = csv.writer(f, delimiter=',', quoting=csv.QUOTE_NONNUMERIC, escapechar=' ')
+            csv_writer.writerow(headers)
 
-    with open(result_file_path, mode='a+', newline='\n') as f:
+    with open(result_file_path, mode='a+', newline='\n', buffering=1) as f:
         csv_writer = csv.writer(f, delimiter=',', quoting=csv.QUOTE_NONNUMERIC, escapechar=' ')
 
         start_time = time.time_ns()
